Module aws_sdk_route53domains::error

Modules

domain_limit_exceeded

See DomainLimitExceeded

duplicate_request

See DuplicateRequest

invalid_input

See InvalidInput

operation_limit_exceeded

See OperationLimitExceeded

tld_rules_violation

See TldRulesViolation

unsupported_tld

See UnsupportedTld

Structs

AcceptDomainTransferFromAnotherAwsAccountError

CancelDomainTransferToAnotherAwsAccountError

CheckDomainAvailabilityError

CheckDomainTransferabilityError

DeleteTagsForDomainError

DisableDomainAutoRenewError

DisableDomainTransferLockError

DomainLimitExceeded

The number of domains has exceeded the allowed threshold for the account.

DuplicateRequest

The request is already in progress for the domain.

EnableDomainAutoRenewError

EnableDomainTransferLockError

GetContactReachabilityStatusError

GetDomainDetailError

GetDomainSuggestionsError

GetOperationDetailError

InvalidInput

The requested item is not acceptable. For example, for APIs that accept a domain name, the request might specify a domain name

ListDomainsError

ListOperationsError

ListTagsForDomainError

OperationLimitExceeded

The number of operations or jobs running exceeded the allowed threshold for the account.

RegisterDomainError

RejectDomainTransferFromAnotherAwsAccountError

RenewDomainError

ResendContactReachabilityEmailError

RetrieveDomainAuthCodeError

TldRulesViolation

The top-level domain does not support this operation.

TransferDomainError

TransferDomainToAnotherAwsAccountError

UnsupportedTld

Amazon Route 53 does not support this top-level domain (TLD).

UpdateDomainContactError

UpdateDomainContactPrivacyError

UpdateDomainNameserversError

UpdateTagsForDomainError

ViewBillingError

Enums

AcceptDomainTransferFromAnotherAwsAccountErrorKind

CancelDomainTransferToAnotherAwsAccountErrorKind

CheckDomainAvailabilityErrorKind

CheckDomainTransferabilityErrorKind

DeleteTagsForDomainErrorKind

DisableDomainAutoRenewErrorKind

DisableDomainTransferLockErrorKind

EnableDomainAutoRenewErrorKind

EnableDomainTransferLockErrorKind

GetContactReachabilityStatusErrorKind

GetDomainDetailErrorKind

GetDomainSuggestionsErrorKind

GetOperationDetailErrorKind

ListDomainsErrorKind

ListOperationsErrorKind

ListTagsForDomainErrorKind

RegisterDomainErrorKind

RejectDomainTransferFromAnotherAwsAccountErrorKind

RenewDomainErrorKind

ResendContactReachabilityEmailErrorKind

RetrieveDomainAuthCodeErrorKind

TransferDomainErrorKind

TransferDomainToAnotherAwsAccountErrorKind

UpdateDomainContactErrorKind

UpdateDomainContactPrivacyErrorKind

UpdateDomainNameserversErrorKind

UpdateTagsForDomainErrorKind

ViewBillingErrorKind