Module aws_sdk_route53domains::input

Modules

accept_domain_transfer_from_another_aws_account_input

See AcceptDomainTransferFromAnotherAwsAccountInput

cancel_domain_transfer_to_another_aws_account_input

See CancelDomainTransferToAnotherAwsAccountInput

check_domain_availability_input

See CheckDomainAvailabilityInput

check_domain_transferability_input

See CheckDomainTransferabilityInput

delete_tags_for_domain_input

See DeleteTagsForDomainInput

disable_domain_auto_renew_input

See DisableDomainAutoRenewInput

disable_domain_transfer_lock_input

See DisableDomainTransferLockInput

enable_domain_auto_renew_input

See EnableDomainAutoRenewInput

enable_domain_transfer_lock_input

See EnableDomainTransferLockInput

get_contact_reachability_status_input

See GetContactReachabilityStatusInput

get_domain_detail_input

See GetDomainDetailInput

get_domain_suggestions_input

See GetDomainSuggestionsInput

get_operation_detail_input

See GetOperationDetailInput

list_domains_input

See ListDomainsInput

list_operations_input

See ListOperationsInput

list_tags_for_domain_input

See ListTagsForDomainInput

register_domain_input

See RegisterDomainInput

reject_domain_transfer_from_another_aws_account_input

See RejectDomainTransferFromAnotherAwsAccountInput

renew_domain_input

See RenewDomainInput

resend_contact_reachability_email_input

See ResendContactReachabilityEmailInput

retrieve_domain_auth_code_input

See RetrieveDomainAuthCodeInput

transfer_domain_input

See TransferDomainInput

transfer_domain_to_another_aws_account_input

See TransferDomainToAnotherAwsAccountInput

update_domain_contact_input

See UpdateDomainContactInput

update_domain_contact_privacy_input

See UpdateDomainContactPrivacyInput

update_domain_nameservers_input

See UpdateDomainNameserversInput

update_tags_for_domain_input

See UpdateTagsForDomainInput

view_billing_input

See ViewBillingInput

Structs

AcceptDomainTransferFromAnotherAwsAccountInput

The AcceptDomainTransferFromAnotherAwsAccount request includes the following elements.

CancelDomainTransferToAnotherAwsAccountInput

The CancelDomainTransferToAnotherAwsAccount request includes the following element.

CheckDomainAvailabilityInput

The CheckDomainAvailability request contains the following elements.

CheckDomainTransferabilityInput

The CheckDomainTransferability request contains the following elements.

DeleteTagsForDomainInput

The DeleteTagsForDomainRequest includes the following elements.

DisableDomainAutoRenewInput

DisableDomainTransferLockInput

The DisableDomainTransferLock request includes the following element.

EnableDomainAutoRenewInput

EnableDomainTransferLockInput

A request to set the transfer lock for the specified domain.

GetContactReachabilityStatusInput

GetDomainDetailInput

The GetDomainDetail request includes the following element.

GetDomainSuggestionsInput

GetOperationDetailInput

The

ListDomainsInput

The ListDomains request includes the following elements.

ListOperationsInput

The ListOperations request includes the following elements.

ListTagsForDomainInput

The ListTagsForDomainRequest includes the following elements.

RegisterDomainInput

The RegisterDomain request includes the following elements.

RejectDomainTransferFromAnotherAwsAccountInput

The RejectDomainTransferFromAnotherAwsAccount request includes the following element.

RenewDomainInput

A RenewDomain request includes the number of years that you want to renew for and the current expiration year.

ResendContactReachabilityEmailInput

RetrieveDomainAuthCodeInput

A request for the authorization code for the specified domain. To transfer a domain to another registrar, you provide

TransferDomainInput

The TransferDomain request includes the following elements.

TransferDomainToAnotherAwsAccountInput

The TransferDomainToAnotherAwsAccount request includes the following elements.

UpdateDomainContactInput

The UpdateDomainContact request includes the following elements.

UpdateDomainContactPrivacyInput

The UpdateDomainContactPrivacy request includes the following elements.

UpdateDomainNameserversInput

Replaces the current set of name servers for the domain with the specified set of name servers.

UpdateTagsForDomainInput

The UpdateTagsForDomainRequest includes the following elements.

ViewBillingInput

The ViewBilling request includes the following elements.