Struct aws_sdk_route53domains::input::RegisterDomainInput

#[non_exhaustive]
pub struct RegisterDomainInput {
    pub domain_name: Option<String>,
    pub idn_lang_code: Option<String>,
    pub duration_in_years: Option<i32>,
    pub auto_renew: Option<bool>,
    pub admin_contact: Option<ContactDetail>,
    pub registrant_contact: Option<ContactDetail>,
    pub tech_contact: Option<ContactDetail>,
    pub privacy_protect_admin_contact: Option<bool>,
    pub privacy_protect_registrant_contact: Option<bool>,
    pub privacy_protect_tech_contact: Option<bool>,
}

The RegisterDomain request includes the following elements.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

domain_name: Option<String>

The domain name that you want to register. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53. For more information, see Formatting Internationalized Domain Names.

idn_lang_code: Option<String>

Reserved for future use.

duration_in_years: Option<i32>

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

auto_renew: Option<bool>

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

admin_contact: Option<ContactDetail>

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

registrant_contact: Option<ContactDetail>

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

tech_contact: Option<ContactDetail>

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

privacy_protect_admin_contact: Option<bool>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

privacy_protect_registrant_contact: Option<bool>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (the domain owner).

Default: true

privacy_protect_tech_contact: Option<bool>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

Implementations

impl RegisterDomainInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<RegisterDomain, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<RegisterDomain>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture RegisterDomainInput