Module aws_sdk_route53domains::operation

Structs

AcceptDomainTransferFromAnotherAwsAccount

Accepts the transfer of a domain from another AWS account to the current AWS account. You initiate a transfer between AWS accounts using

CancelDomainTransferToAnotherAwsAccount

Cancels the transfer of a domain from the current AWS account to another AWS account. You initiate a transfer between AWS accounts using

CheckDomainAvailability

This operation checks the availability of one domain name. Note that if the availability status of a domain is pending, you must

CheckDomainTransferability

Checks whether a domain name can be transferred to Amazon Route 53.

DeleteTagsForDomain

This operation deletes the specified tags for a domain.

DisableDomainAutoRenew

This operation disables automatic renewal of domain registration for the specified domain.

DisableDomainTransferLock

This operation removes the transfer lock on the domain (specifically the

EnableDomainAutoRenew

This operation configures Amazon Route 53 to automatically renew the specified domain before the domain registration expires.

EnableDomainTransferLock

This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status)

GetContactReachabilityStatus

For operations that require confirmation that the email address for the registrant contact is valid,

GetDomainDetail

This operation returns detailed information about a specified domain that is associated with the current AWS account.

GetDomainSuggestions

The GetDomainSuggestions operation returns a list of suggested domain names.

GetOperationDetail

This operation returns the current status of an operation that is not completed.

ListDomains

This operation returns all the domain names registered with Amazon Route 53 for the current AWS account.

ListOperations

Returns information about all of the operations that return an operation ID and that have ever been

ListTagsForDomain

This operation returns all of the tags that are associated with the specified domain.

RegisterDomain

This operation registers a domain. Domains are registered either by Amazon Registrar (for .com, .net, and .org domains) or by

RejectDomainTransferFromAnotherAwsAccount

Rejects the transfer of a domain from another AWS account to the current AWS account. You initiate a transfer between AWS accounts using

RenewDomain

This operation renews a domain for the specified number of years. The cost of renewing your domain is billed to your AWS account.

ResendContactReachabilityEmail

For operations that require confirmation that the email address for the registrant contact is valid,

RetrieveDomainAuthCode

This operation returns the AuthCode for the domain. To transfer a domain to another registrar, you provide this value to the new registrar.

TransferDomain

Transfers a domain from another registrar to Amazon Route 53. When the transfer is complete, the domain is registered either with

TransferDomainToAnotherAwsAccount

Transfers a domain from the current AWS account to another AWS account. Note the following:

UpdateDomainContact

This operation updates the contact information for a particular domain. You must specify information for at least one contact:

UpdateDomainContactPrivacy

This operation updates the specified domain contact's privacy setting. When privacy protection is enabled,

UpdateDomainNameservers

This operation replaces the current set of name servers for the domain with the specified set of name servers.

UpdateTagsForDomain

This operation adds or updates tags for a specified domain.

ViewBilling

Returns all the domain-related billing records for the current AWS account for a specified period